+++ Rendering/FrameBuffer.cs
@@ -7,108 +7,21 @@ using System.Reflection;
 
 namespace DiscordGameEngine.Rendering
 {
-
-    public enum PixelRenderMode
-    { 
-        ALPHA_BLENDING=0,
-        NORMAL=1,
-        REPLACE=2
-    }
-
-    public class FrameBuffer
+    public class FrameBuffer : PixelBuffer
     {
         private static string imageBuffersPath = System.IO.Path.GetDirectoryName(Assembly.GetEntryAssembly().Location) + "\\imageBuffers\\";
         private static System.Drawing.Imaging.ImageFormat imageBufferSaveFormat = System.Drawing.Imaging.ImageFormat.Png;
         private static Random randomGen = new Random();
         private static List<string> imageBufferIDs = new List<string>();
 
-        public Size size { get { return size; } }
-        private Size _size;
-        public Color clearColor;
-        public Bitmap buffer;
-        public PixelRenderMode pixelRenderMode;
-
         public event EventHandler OnResize;
-
         private string imageBufferID;
 
-        public FrameBuffer(Size size, Color clearColor)
+        public FrameBuffer(Size size, Color clearColor) : base(size, clearColor)
         {
-            this._size = size;
-            this.clearColor = clearColor;
-            this.pixelRenderMode = PixelRenderMode.NORMAL;
             CheckForImageBuffersDir();
             this.imageBufferID = GetNewImageBufferID();
-            CreateBuffer();
-            Clear();
             Render();
-        }
-
-        private Color AlphaBlend(Color c1, Color c2)
-        {
-            float a1 = c1.A / 255f;
-            float a2 = c2.A / 255f;
-            c1 = Color.FromArgb((int)(c1.R * a1), (int)(c1.G * a1), (int)(c1.B * a1));
-            c2 = Color.FromArgb((int)(c2.R * a2), (int)(c2.G * a2), (int)(c2.B * a2));
-            return Color.FromArgb(Math.Min(255, (int) ((c1.A + c2.A) * (1 - a1))), Math.Min(255, (int) ((c1.R + c2.R) * (1 - a1))), Math.Min(255, (int) ((c1.G + c2.G) * (1 - a1))), Math.Min(255, (int) ((c1.B + c2.B) * (1 - a1))));
-        }
-
-        private Color ToOpaqueColor(Color color)
-        {
-            return Color.FromArgb(color.R, color.G, color.B);
-        }
-
-        public void Clear()
-        {
-            for (int x = 0; x < _size.Width; x++)
-            {
-                for (int y = 0; y < _size.Height; y++)
-                {
-                    buffer.SetPixel(x, y, clearColor);
-                }
-            }
-        }
-
-        public void Draw(int x, int y, Color color)
-        {
-            if (pixelRenderMode == PixelRenderMode.ALPHA_BLENDING)
-            {
-                color = AlphaBlend(color, buffer.GetPixel(x, y));
-                return; 
-            }
-            color = ToOpaqueColor(color);
-            buffer.SetPixel(x, y, color);
-        }
-
-        public void DrawRect(int x, int y, int sizeX, int sizeY, Color color)
-        {
-            int x1 = x + sizeX;
-            int y1 = y + sizeY;
-            if (pixelRenderMode == PixelRenderMode.ALPHA_BLENDING)
-            {
-                for (int i = x; i < x1; i++)
-                {
-                    for (int j = y; j < y1; j++)
-                    {
-                        buffer.SetPixel(i, j, AlphaBlend(color, buffer.GetPixel(i, j)));
-                    }
-                }
-            }
-            else
-            {
-                color = ToOpaqueColor(color);
-                for (int i = x; i < x1; i++)
-                {
-                    for (int j = y; j < y1; j++)
-                    {
-                        buffer.SetPixel(i, j, color);
-                    }
-                }
-            }
-        }
-
-        public void DrawLine()
-        {
         }
 
         /// <summary>
@@ -124,13 +37,6 @@ namespace DiscordGameEngine.Rendering
             context.Channel.SendFileAsync(imageBuffersPath + imageBufferID);
         }
 
-        public void SetSize(Size newSize)
-        {
-            _size = newSize;
-            ResizeBuffer();
-            OnResize?.Invoke(this, new EventArgs());
-        }
-
         private void SaveBufferToDisk()
         {
             buffer.Save(imageBuffersPath + imageBufferID, imageBufferSaveFormat);
@@ -138,18 +44,19 @@ namespace DiscordGameEngine.Rendering
 
         private void CreateBuffer()
         {
-            buffer = new Bitmap(_size.Width, _size.Height);
+            buffer = new Bitmap(size.Width, size.Height);
         }
 
-        private void ResizeBuffer()
+        public override void SetSize(Size newSize)
         {
-            buffer = new Bitmap(buffer, size);
+            base.SetSize(newSize);
+            OnResize?.Invoke(this, new EventArgs());
         }
 
         /// <summary>
         /// Call to free RAM and Disk space if this FrameBuffer is no longer used -> using it after calling this method might create bugs
         /// </summary>
-        public void Dispose()
+        public override void Dispose()
         {
             buffer.Dispose();
             ClearStoredImageBuffer();
@@ -169,7 +76,7 @@ namespace DiscordGameEngine.Rendering
             string res;
             do
             {
-                res = "imageBuffer" + randomGen.Next(1000000, 10000000) + '.' + imageBufferSaveFormat.ToString().ToLower();
+                res = "imageBuffer" + randomGen.Next(1000000, 10000000 - 1) + '.' + imageBufferSaveFormat.ToString().ToLower();
             } while (imageBufferIDs.Contains(res));
             imageBufferIDs.Add(res);
             return res;